requirements.txt
guildai==0.9.0
pandas~=2.1.0
fastapi~=0.103.1
FastAPI-SQLAlchemy~=0.2.1
uvicorn==0.23.2
jinja2
pyod==1.1.0
joblib~=1.3.2
numpy~=1.24.4
boto3~=1.28.50
botocore~=1.31.50
scikit-learn~=1.3.0
PyYAML~=6.0.1
PyJWT~=2.8.0
pydantic~=2.3.0
passlib~=1.7.4
SQLAlchemy~=2.0.23
paho-mqtt~=1.6.1
python-dotenv~=1.0.0
certifi~=2023.7.22
pika~=1.3.2
psycopg2~=2.9.9
httpx~=0.25.2

cryptography~=41.0.7
alembic~=1.13.0
starlette~=0.27.0